[conf] -Fix: compiler warning
git-svn-id: svn://repo.hu/pcb-rnd/trunk@1911 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/conf.c
@@ -428,7 +428,7 @@ int conf_merge_patch_recurse(lht_node_t *sect, int default_prio, conf_policy_t d
 	return res;
 }
 
-int conf_merge_patch(lht_node_t *root, int gprio)
+int conf_merge_patch(lht_node_t *root, long gprio)
 {
 	conf_policy_t gpolicy = POL_OVERWRITE;
 	const char *ps;
